requirements.txt
numpy>=1.26.4
PyOpenGL>=3.1.7
PySide6>=6.6.1
pytest>=7.4.2  # for testing
black>=24.1.1  # for code formatting